+++ pydantic/utils.py
@@ -1,7 +1,9 @@
 import re
 from contextlib import contextmanager
+from enum import Enum
 from importlib import import_module
-from typing import Tuple, _TypingBase
+from textwrap import dedent
+from typing import Tuple
 
 from . import errors
 
@@ -9,6 +11,11 @@ try:
     import email_validator
 except ImportError:
     email_validator = None
+
+try:
+    from typing import _TypingBase as typing_base
+except ImportError:
+    from typing import _Final as typing_base
 
 
 PRETTY_REGEX = re.compile(r'([\w ]*?) *<(.*)> *')
@@ -115,8 +122,16 @@ def truncate(v, *, max_len=80):
 
 
 def display_as_type(v):
-    if not isinstance(v, _TypingBase) and not isinstance(v, type):
+    if not isinstance(v, typing_base) and not isinstance(v, type):
         v = type(v)
+
+    if isinstance(v, type) and issubclass(v, Enum):
+        if issubclass(v, int):
+            return 'int'
+        elif issubclass(v, str):
+            return 'str'
+        else:
+            return 'enum'
 
     try:
         return v.__name__
@@ -131,3 +146,7 @@ def change_exception(raise_exc, *except_types):
         yield
     except except_types as e:
         raise raise_exc from e
+
+
+def clean_docstring(d):
+    return dedent(d).strip(' \r\n\t')