+++ tests/test_main.py
@@ -54,9 +54,10 @@ def test_ultra_simple_failed():
 def test_ultra_simple_repr():
     m = UltraSimpleModel(a=10.2)
     assert repr(m) == '<UltraSimpleModel a=10.2 b=10>'
-    assert repr(m.fields['a']) == ("<Field a: type='float', required=True, "
-                                   "validators=['float_validator']>")
+    assert repr(m.fields['a']) == "<Field(a type=float required)>"
     assert dict(m) == {'a': 10.2, 'b': 10}
+    assert m.dict() == {'a': 10.2, 'b': 10}
+    assert m.json() == '{"a": 10.2, "b": 10}'
 
 
 def test_str_truncate():