+++ HISTORY.rst
@@ -3,10 +3,17 @@
 History
 -------
 
-v0.11.0 (2018-XX-XX)
+v0.11.1 (2018-07-01)
+....................
+* support Python 3.7
+
+v0.11.0 (2018-06-28)
 ....................
 * make ``list``, ``tuple`` and ``set`` types stricter #86
 * **breaking change**: remove msgpack parsing #201
+* add ``FilePath`` and ``DirectoryPath`` types #10
+* model schema generation #190
+* JSON serialisation of models and schemas #133
 
 v0.10.0 (2018-06-11)
 ....................